backend/pyproject.toml
[project]
name = "aq-trading"
version = "0.1.0"
description = "Algorithmic trading system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "httpx>=0.26.0",
    "pyyaml>=6.0",
    "aiosmtplib>=3.0.0",
    "jsonpatch>=1.33",
    "futu-api>=9.0.0",
    "pydantic-yaml>=1.3.0",
    "APScheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "aiosqlite>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
markers = [
    "timescaledb: marks tests as requiring TimescaleDB (deselect with '-m \"not timescaledb\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # assert used (ok in tests)
    "B008",   # function call in default argument (FastAPI Depends)
    "UP017",  # datetime.UTC requires Python 3.11+; keep timezone.utc for 3.10 compatibility
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "S105", "S106", "F841"]  # Allow asserts, hardcoded passwords, unused vars in tests
"src/broker/paper_broker.py" = ["S311"]  # Pseudo-random OK for paper trading simulation

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
